+++ src/gui/mainwindow.cpp
@@ -2509,6 +2509,9 @@ void MainWindow::loadSettings(QSettings &settings, AppConfig &appConfig)
     m_trayMenu->setNumberSearchEnabled(m_sharedData->numberSearch);
     m_menu->setNumberSearchEnabled(m_sharedData->numberSearch);
 
+    m_trayMenu->setRowIndexFromOne(m_sharedData->rowIndexFromOne);
+    m_menu->setRowIndexFromOne(m_sharedData->rowIndexFromOne);
+
     m_options.transparency = appConfig.option<Config::transparency>();
     m_options.transparencyFocused = appConfig.option<Config::transparency_focused>();
     updateWindowTransparency();