+++ src/gui/traymenu.cpp
@@ -101,10 +101,11 @@ void TrayMenu::addClipboardItemAction(const QVariantMap &data, bool showImages)
     QString format;
 
     // Add number key hint.
-    if (m_clipboardItemActionCount < 10) {
+    const int rowNumber = m_clipboardItemActionCount + static_cast<int>(m_rowIndexFromOne);
+    if (rowNumber < 10) {
         format = tr("&%1. %2",
                     "Key hint (number shortcut) for items in tray menu (%1 is number, %2 is item label)")
-                .arg(m_clipboardItemActionCount);
+                .arg(rowNumber);
     }
 
     m_clipboardItemActionCount++;